requirements.txt
flask
gunicorn
requests
pyodbc
python-dotenv